move old foreign API to module
